notifications/requirements.txt
# Transactional email (https://resend.com)
resend>=2.0.0
python-dotenv>=1.0.0

# Dev upload API (roster_upload_app)
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
